just put everything in __init__.py
