--- src/h3/api/numpy_int/_api_template.py
@@ -1,1 +1,0 @@
-../basic_int/_api_template.py
