+++ tests/test_api_bindings_match.py
@@ -6,10 +6,10 @@ def test_api_copy_match():
     import h3.api.numpy_int
 
     apis = [
-        h3.api.basic_int._api_template,
-        h3.api.basic_str._api_template,
-        h3.api.memview_int._api_template,
-        h3.api.numpy_int._api_template,
+        h3.api.basic_int,
+        h3.api.basic_str,
+        h3.api.memview_int,
+        h3.api.numpy_int,
     ]
 
     api_set = {inspect.getsource(api) for api in apis}
